Adding Pause and seed change system.

+++ main.py
@@ -3,6 +3,31 @@ import random
 from time import sleep,time
 from Block import Block
 from ImageBackGround import *
+from tkinter import *
+from myCheckbox import *
+def confirmSeed():
+    global eSeed
+    global r
+    global isRandSeed
+    print(randseedCheck.getVar())
+    isRandSeed =not( randseedCheck.getVar())
+    r = int(eSeed.get())
+    random.seed(int(eSeed.get()))
+    seedWindow.withdraw()
+    main.quit()
+    restart(r)
+isRandSeed = True
+main = Tk()
+seedWindow = Toplevel()
+label = Label(seedWindow,text = "Enter seed:")
+eSeed = Entry(seedWindow)
+randseedCheck = mycheckbox(seedWindow,"Randomize seed",True,False)
+bSeed = Button(seedWindow,text = "confirm",command = confirmSeed)
+label.pack()
+eSeed.pack()
+randseedCheck.pack()
+bSeed.pack()
+main.withdraw()
 pygame.init()
 isDebug = True
 transition = pygame.display.set_mode((400,400))
@@ -29,9 +54,12 @@ RED = (255, 0, 0)
 BLUE = (0,0,153)
 GREEN = (0,255,0)
 PLAYER_SPEED = 4
+pPlayer_SPEED = 4
 SCROLL_SPEED = 4
+pSCROLL_SPEED = 4
 r = random.randint(0,1000)
 random.seed(r)
+
 def makeCol(x,amountCols):
     newBlocks = []
     blockPlacements = []
@@ -44,7 +72,30 @@ def makeCol(x,amountCols):
     for i in range(0,len(blockPlacements)):
         newBlocks.append(Block(BLUE,screen,blockPlacements[i][0]*32,blockPlacements[i][1]*32))
     return newBlocks
-def restart():
+
+
+def pause(screen):
+    global SCROLL_SPEED
+    global PLAYER_SPEED
+    global pSCROLL_SPEED
+    global pPlayer_SPEED
+    global isPause
+    global pauseTime
+    pauseTime = time()
+    isPause = not isPause
+    if isPause:
+        pSCROLL_SPEED = SCROLL_SPEED
+        pPlayer_SPEED = PLAYER_SPEED
+        PLAYER_SPEED = 0
+        SCROLL_SPEED = 0
+    if not isPause:
+        PLAYER_SPEED = pPlayer_SPEED
+        SCROLL_SPEED = pSCROLL_SPEED
+        screen.fill(BLACK)
+        downBlue.draw(screen)
+
+
+def restart(seed = None):
     global player
     global player2
     global blocks
@@ -56,9 +107,17 @@ def restart():
     global SCROLL_SPEED
     global players
     global score
+    global isPause
+    global screen
+    global isRandSeed
+    if isPause:
+        pause(screen)
     score = 0
     SCROLL_SPEED = 4
-    r = random.randint(0,1000)
+    if seed is not None:
+        r = seed
+    elif isRandSeed:
+        r = random.randint(0,1000)
     random.seed(r)
     counter = 0
     start = time()
@@ -76,7 +135,7 @@ def restart():
     upBlue.draw(screen)
     downBlue.draw(screen)
 
-
+pauseTime = 0
 failFont = pygame.font.SysFont('Arial', 52)
 tFailed = failFont.render('GAME OVER!', True, (255, 0, 0))
 TOTAL_H = 384
@@ -86,8 +145,8 @@ players = []
 player = Block(RED, screen, 0,32,28,28,True,players)
 player2 = Block(GREEN,screen,0,32,28,28,True,players)
 players.append(player)
-print(player.groups())
 isPlayerTwo = False
+isPause = False
 running = True
 isDead = False
 start = time()
@@ -98,19 +157,24 @@ downBlue = Block(BLUE,screen, 0,TOTAL_H - 32,TOTAL_W)
 upBlue.draw(screen)
 downBlue.draw(screen)
 timefont = pygame.font.SysFont('Consolas', 13)
+tPause = failFont.render("PAUSE",True,RED)
 counter = 0
 score = 0
 while running:
+    if isPause:
+        screen.blit(tPause,(150,200))
+        pygame.display.flip()
     keys = pygame.key.get_pressed()
     player.draw(screen)
-    SCROLL_SPEED += 0.001
+    if not isPause:
+        SCROLL_SPEED += 0.001
     pygame.time.delay(16)
     if not isDead:
         if counter % 64 == 0:
             blocks.extend(makeCol(33,8))
         upBlue.draw(screen)
         tScore = timefont.render("Score: " + str(round(score,0)), True, RED)
-        tTime = timefont.render("seed: "+str(r) + ", Scroll Speed: "+ str(round(SCROLL_SPEED,3)) + ", Time: " + str(round(time()-start,3)) , True, (255, 0, 0))
+        tTime = timefont.render("seed: "+str(r) + ", Scroll Speed: "+ str(round(SCROLL_SPEED,3)) , True, (255, 0, 0))
         screen.blit(tTime,(10,10))
         screen.blit(tScore,(600,10))
         for i in blocks:
@@ -122,6 +186,13 @@ while running:
                     screen.blit(tFailed,(200,150))
                     isDead = True
         pygame.display.flip()
+        if keys[pygame.K_s] and isPause:
+            print("Opening main")
+            seedWindow.deiconify()
+            main.mainloop()
+
+        if keys[pygame.K_r] and time()-pauseTime > 1:
+            pause(screen)
         if keys[pygame.K_UP]:
             player.update(0,-PLAYER_SPEED)
         if keys[pygame.K_DOWN]:
@@ -130,15 +201,13 @@ while running:
             player.update(PLAYER_SPEED,0)
         if keys[pygame.K_LEFT]:
             player.update(-PLAYER_SPEED,0)
-        if pygame.key.get_pressed()[pygame.K_p] and time() - time_since_change > 1:
+        if pygame.key.get_pressed()[pygame.K_p] and time() - time_since_change > 1 and (not isPause):
             time_since_change = time()
             if isPlayerTwo:
-                print(players)
                 isPlayerTwo = False
                 # players.remove(player2)
                 player2.kill()
             else:
-                print(players)
                 player2.x = 0
                 player2.y = 32
                 player2.updateShape()
@@ -160,9 +229,9 @@ while running:
             running = False
     if keys[pygame.K_ESCAPE]:
         running = False
-    if keys[pygame.K_SPACE]:
+    if keys[pygame.K_SPACE] and not isPause:
         restart()
-    if not isDead:
+    if not isDead and not isPause:
         counter += 1
-        score += len(blocks)/10
+        score += len(blocks)
